Routes protection

+++ routes/web.php
@@ -21,34 +21,34 @@ use App\Models\Listing;
 Route::get('/', [ListingController::class, 'index']);
 
 //Create form 
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store listings
-Route::post('/listings',[ListingController::class, 'store']);
+Route::post('/listings',[ListingController::class, 'store'])->middleware('auth');
 
 //Edit form
-Route::get('/listings/{listing}/edit',[ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class, 'edit'])->middleware('auth');
 
 //Update listing
-Route::put('/listings/{listing}',[ListingController::class, 'update']);
+Route::put('/listings/{listing}',[ListingController::class, 'update'])->middleware('auth');
 
 //Delete listing
-Route::delete('/listings/{listing}',[ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class, 'destroy'])->middleware('auth');
 
 //Single listings
 Route::get('/listings/{listing}',[ListingController::class, 'show']);
 
 //Registration form
-Route::get('/register',[UserController::class, 'register']);
+Route::get('/register',[UserController::class, 'register'])->middleware('guest');
 
 //Create new user
 Route::post('/users',[UserController::class, 'store']);
 
 //Log user out
-Route::post('/logout',[UserController::class, 'logout']);
+Route::post('/logout',[UserController::class, 'logout'])->middleware('auth');
 
 //Show login form
-Route::get('/login',[UserController::class, 'login']);
+Route::get('/login',[UserController::class, 'login'])->name('login')->middleware('guest');
 
 //Login user
 Route::get('/users/authenticate',[UserController::class, 'authenticate']);
